Change Action and EventListener speeds to f64 for fractional speed support
This implements JavaScript's fractional speed system where speeds are adjusted by small fractions to break ties deterministically. Changed:
- Action speeds (MoveAction, SwitchAction, TeamAction, PokemonAction) from i32 to f64
- EventListener speed from Option<i32> to Option<f64>
- PriorityItem speed from i32 to f64
- Implemented fractional speed adjustment for SwitchIn event handlers
- Updated all comparison functions to use total_cmp() for f64

+++ src/battle.rs
@@ -211,7 +211,7 @@ pub struct EventListener {
     /// Effect order (for hazards and abilities with same priority)
     pub effect_order: Option<i32>,
     /// Speed stat (for speed-based sorting)
-    pub speed: Option<i32>,
+    pub speed: Option<f64>,
 }
 
 /// Effect type - matches JavaScript effectType
@@ -593,7 +593,7 @@ impl Battle {
 pub struct PriorityItem {
     pub order: Option<i32>,
     pub priority: i32,
-    pub speed: i32,
+    pub speed: f64,  // Changed from i32 to f64 to match JavaScript's fractional speeds
     pub sub_order: i32,
     pub effect_order: i32,
     pub index: usize,
